cmd: Reorganize dev cmds and dev mode check
